pyproject.toml
[tool.poetry]
name = "tts-api"
version = "0.1.0"
description = ""
authors = ["Kuldeep <codysharma01@gmailc.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
Django = "^5.0.6"
django-rest-framework = "^0.1.0"
tts = "^0.22.0"
celery = "^5.4.0"
redis = "^5.0.6"
boto3 = "^1.34.127"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
